requirements.txt
google-generativeai
PyMuPDF
python-dotenv
